[cmake][fix][msvc] Add support for users with spaces in their pathnames, by enclosing the `/DEF` link flag value in quotes.
Until now, if your path to the TaikoArcadeLoader repository contained spaces and you tried building bnusio, it would fail due to a linker error that looks something like this (from a real-world example):

> LINK : fatal error LNK1181: cannot open input file 'Hinkle\Documents\TaikoArcadeLoader\exports.def' [C:\Users\Christian Hinkle\Documents\TaikoArcadeLoader\build\bnusio.vcxproj]

Notice that it failed to open the "exports.def" file because the full file path was not used. It ignored everything preceeding the last space in the string. To fix this, we wrap the file path value in quotes, to indicate that the entire string (including spaces) should be treated as a single argument.